requirements.txt
abstra==3.3.0
pillow==10.4.0
fpdf